+++ web/app/(extension-layout)/extension/layout.tsx
@@ -1,5 +1,6 @@
 import RemoteModuleProvider from "@/components/providers/remote-module-provider";
 import { ReactNode, Suspense } from "react";
+import { Analytics } from "@vercel/analytics/next";
 
 export default function ExtensionLayout({ children }: { children: ReactNode }) {
   return (
@@ -10,6 +11,7 @@ export default function ExtensionLayout({ children }: { children: ReactNode }) {
           width: "100vw",
         }}
       >
+        <Analytics />
         <RemoteModuleProvider isPreventingCSS={false}>
           <Suspense>{children}</Suspense>
         </RemoteModuleProvider>