+++ web/app/(main-layout)/layout.tsx
@@ -8,6 +8,8 @@ import CapacitorProvider from "@/components/providers/capacitor-provider";
 import RemoteModuleProvider from "@/components/providers/remote-module-provider";
 import InterModuleCommunicationProvider from "@/components/providers/imc-provider";
 import Nav from "@/components/interface/nav";
+import { Suspense } from "react";
+import { Analytics } from "@vercel/analytics/next";
 
 export const metadata: Metadata = {
   title: "Pulse Editor",
@@ -22,18 +24,21 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={`h-[100dvh] w-[100dvw] antialiased`}>
-        <CapacitorProvider>
-          <WrappedHeroUIProvider>
-            <EditorContextProvider>
-              <InterModuleCommunicationProvider>
-                <RemoteModuleProvider isPreventingCSS={true}>
-                  <Toaster />
-                  <Nav>{children}</Nav>
-                </RemoteModuleProvider>
-              </InterModuleCommunicationProvider>
-            </EditorContextProvider>
-          </WrappedHeroUIProvider>
-        </CapacitorProvider>
+        <Analytics />
+        <Suspense>
+          <CapacitorProvider>
+            <WrappedHeroUIProvider>
+              <EditorContextProvider>
+                <InterModuleCommunicationProvider>
+                  <RemoteModuleProvider isPreventingCSS={true}>
+                    <Toaster />
+                    <Nav>{children}</Nav>
+                  </RemoteModuleProvider>
+                </InterModuleCommunicationProvider>
+              </EditorContextProvider>
+            </WrappedHeroUIProvider>
+          </CapacitorProvider>
+        </Suspense>
       </body>
     </html>
   );