+++ web/components/extension/extension-list.tsx
@@ -1,6 +1,7 @@
 import { Extension } from "@/lib/types";
 import Loading from "../interface/loading";
 import ExtensionPreview from "./preview";
+import { compare } from "semver";
 
 export default function ExtensionList({
   extensions,
@@ -11,11 +12,24 @@ export default function ExtensionList({
   isLoading: boolean;
   showInstalledChip: boolean;
 }) {
+  // Group extensions by name
+  const groupedExtensions = extensions.reduce(
+    (acc: Map<string, Extension[]>, ext) => {
+      const key = ext.config.id;
+      if (!acc.has(key)) {
+        acc.set(key, []);
+      }
+      acc.get(key)?.push(ext);
+      return acc;
+    },
+    new Map<string, Extension[]>(),
+  );
+
   return (
     <>
       {isLoading ? (
         <Loading />
-      ) : extensions.length === 0 ? (
+      ) : groupedExtensions.size === 0 ? (
         <div className="w-full space-y-2">
           <p className="text-center text-lg">No extensions found</p>
           <p>
@@ -25,13 +39,22 @@ export default function ExtensionList({
         </div>
       ) : (
         <div className="grid grid-cols-2 gap-1">
-          {extensions.map((ext) => (
-            <ExtensionPreview
-              extension={ext}
-              key={ext.config.id + ext.config.version}
-              showInstalledChip={showInstalledChip}
-            />
-          ))}
+          {Array.from(groupedExtensions.entries()).map(([name, extGroup]) => {
+            // Take the latest version of each extension group
+            const latestVersion = extGroup.reduce((latest, current) => {
+              return compare(current.config.version, latest.config.version) > 0
+                ? current
+                : latest;
+            }, extGroup[0]);
+
+            return (
+              <ExtensionPreview
+                key={name}
+                extension={latestVersion}
+                showInstalledChip={showInstalledChip}
+              />
+            );
+          })}
         </div>
       )}
     </>