+++ web/components/interface/extension-suggestion-overlay.tsx
@@ -3,9 +3,9 @@ import ExtensionPreview from "../extension/preview";
 export default function ExtensionSuggestionOverlay() {
   return (
     <div className="absolute top-0 left-0 h-full w-full backdrop-blur-xs backdrop-invert-25">
-      <div className="flex h-full w-full grid-rows-[max-content_auto] flex-col items-center pb-10 pt-4">
+      <div className="flex h-full w-full grid-rows-[max-content_auto] flex-col items-center pt-4 pb-10">
         <div className="bg-content1 responsive-content h-20 rounded-lg"></div>
-        <div className="h-full flex flex-col justify-center pb-20">
+        <div className="flex h-full flex-col justify-center pb-20">
           <div className="responsive-content flex gap-2">
             <ExtensionPreview
               showInstalledChip
@@ -15,9 +15,10 @@ export default function ExtensionSuggestionOverlay() {
                   displayName: "Extension Name",
                   description: "Extension Description",
                   version: "1.0.0",
+                  visibility: "public",
                 },
                 isEnabled: false,
-                remoteOrigin: "https://cdn.pulse-editor.com/extension",
+                remoteOrigin: `${process.env.NEXT_PUBLIC_CDN_URL}/${process.env.NEXT_PUBLIC_STORAGE_CONTAINER}`,
               }}
             />
             <ExtensionPreview
@@ -28,9 +29,10 @@ export default function ExtensionSuggestionOverlay() {
                   displayName: "Extension Name",
                   description: "Extension Description",
                   version: "1.0.0",
+                  visibility: "public",
                 },
                 isEnabled: false,
-                remoteOrigin: "https://cdn.pulse-editor.com/extension",
+                remoteOrigin: `${process.env.NEXT_PUBLIC_CDN_URL}/${process.env.NEXT_PUBLIC_STORAGE_CONTAINER}`,
               }}
             />
             <ExtensionPreview
@@ -41,9 +43,10 @@ export default function ExtensionSuggestionOverlay() {
                   displayName: "Extension Name",
                   description: "Extension Description",
                   version: "1.0.0",
+                  visibility: "public",
                 },
                 isEnabled: false,
-                remoteOrigin: "https://cdn.pulse-editor.com/extension",
+                remoteOrigin: `${process.env.NEXT_PUBLIC_CDN_URL}/${process.env.NEXT_PUBLIC_STORAGE_CONTAINER}`,
               }}
             />
           </div>