+++ web/lib/hooks/use-auth.ts
@@ -1,15 +1,14 @@
+"use client";
+
 import { useContext } from "react";
 import { Session } from "../types";
 import useSWR from "swr";
 import { EditorContext } from "@/components/providers/editor-context-provider";
 
-const authUrl = "https://pulse-editor.com";
-// const authUrl = "https://localhost:8080";
-
 export function useAuth() {
   // --- Auth ---
   const { data: session, isLoading } = useSWR<Session | undefined>(
-    `${authUrl}/api/auth/session`,
+    `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/auth/session`,
     async (url: string) => {
       const res = await fetch(url, {
         credentials: "include",
@@ -34,7 +33,9 @@ export function useAuth() {
       return;
     }
 
-    const url = new URL(`${authUrl}/api/auth/signin`);
+    const url = new URL(
+      `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/auth/signin`,
+    );
     url.searchParams.set("callbackUrl", window.location.href);
 
     window.location.href = url.toString();
@@ -46,7 +47,9 @@ export function useAuth() {
       return;
     }
 
-    const url = new URL(`${authUrl}/api/auth/signout`);
+    const url = new URL(
+      `${process.env.NEXT_PUBLIC_BACKEND_URL}/api/auth/signout`,
+    );
     url.searchParams.set("callbackUrl", window.location.href);
     window.location.href = url.toString();
   }